requirements.txt
-e .
pytest
aws_cdk.aws_codecommit
aws_cdk.aws_codepipeline
aws_cdk.aws_codebuild
aws_cdk.aws_cloudwatch
aws_cdk.aws_codepipeline_actions
aws_cdk.pipelines
aws_cdk.aws_ecr 
aws_cdk.aws_iam
aws_cdk.core